Share code with the existing implementation
